// src/platform/packages/shared/kbn-lens-common/types.d.ts
export type IndexPatternField = FieldSpec & {
    displayName: string;
    aggregationRestrictions?: Partial<IndexPatternAggRestrictions>;
    /**
     * Map of fields which can be used, but may fail partially (ranked lower than others)
     */
    partiallyApplicableFunctions?: Partial<Record<string, boolean>>;
    timeSeriesMetric?: 'histogram' | 'summary' | 'gauge' | 'counter' | 'position';
    timeSeriesRollup?: boolean;
    meta?: boolean;
    runtime?: boolean;
};
export interface DateRange {
    fromDate: string;
    toDate: string;
}
interface PersistableFilterMeta extends FilterMeta {
    indexRefName?: string;
}
export interface PersistableFilter extends Filter {
    meta: PersistableFilterMeta;
}
export type SortingHint = string;
export type ValueLabelConfig = 'hide' | 'show';
export interface IndexPattern {
    getFormatterForField(// used extensively in lens
    sourceField: string): unknown;
    id: string;
    fields: IndexPatternField[];
    getFieldByName(name: string): IndexPatternField | undefined;
    title: string;
    name?: string;
    timeFieldName?: string;
    fieldFormatMap?: Record<string, {
        id: string;
        params: FieldFormatParams;
    }>;
    hasRestrictions: boolean;
    spec: DataViewSpec;
    isPersisted: boolean;
}
// ...
export interface ValueFormatConfig {
    id: string;
    params?: {
        decimals: number;
        suffix?: string;
        compact?: boolean;
        pattern?: string;
        fromUnit?: string;
        toUnit?: string;
    };
}