// src/platform/packages/shared/kbn-lens-common/datasources/types.d.ts
export type OperationType = string;
export interface BaseIndexPatternColumn extends Operation {
    operationType: string;
    customLabel?: boolean;
    timeScale?: TimeScaleUnit;
    filter?: Query;
    reducedTimeRange?: string;
    timeShift?: string;
}
export interface FormattedIndexPatternColumn extends BaseIndexPatternColumn {
    params?: {
        format?: ValueFormatConfig;
    };
}
export interface FieldBasedIndexPatternColumn extends BaseIndexPatternColumn {
    sourceField: string;
}
export interface ReferenceBasedIndexPatternColumn extends FormattedIndexPatternColumn {
    references: string[];
}
export type GenericIndexPatternColumn = BaseIndexPatternColumn | FieldBasedIndexPatternColumn | ReferenceBasedIndexPatternColumn;
// ...
export interface IncompleteColumn {
    operationType?: OperationType;
    sourceField?: string;
}
export interface FormBasedLayer {
    columnOrder: string[];
    columns: Record<string, GenericIndexPatternColumn>;
    indexPatternId: string;
    linkToLayers?: string[];
    incompleteColumns?: Record<string, IncompleteColumn | undefined>;
    sampling?: number;
    ignoreGlobalFilters?: boolean;
}